requirements.txt
streamlit==1.32.2
pandas==2.2.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.28
python-dotenv==1.0.1
plotly_express==0.4.0